# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OrderPing - Food court order notification system API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stall {
  id: string;
  name: string;
  mallName: string;
  slug: string;
  email: string;
  createdAt: string;
  currentlyServing?: string | null;
}

export interface CreateStallBody {
  name: string;
  mallName: string;
  email: string;
  password: string;
  slug?: string;
}

export interface VerifyPasswordBody {
  password: string;
}

export interface VerifyPasswordResult {
  success: boolean;
  stallId: string;
  stallName: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  waiting: "waiting",
  ready: "ready",
  completed: "completed",
} as const;

export interface Order {
  id: string;
  stallId: string;
  stallSlug: string;
  receiptNumber: string;
  status: OrderStatus;
  createdAt: string;
  readyAt?: string | null;
  completedAt?: string | null;
  nudgeCount: number;
  lastNudgeAt?: string | null;
}

export interface CreateOrderBody {
  receiptNumber: string;
}

export interface QueueStatus {
  stallName: string;
  mallName: string;
  currentlyServing?: string | null;
  waitingCount: number;
  recentlyServed: string[];
}

export interface HourlyCount {
  hour: string;
  count: number;
}

export interface StallAnalytics {
  ordersToday: number;
  avgWaitTimeMinutes: number;
  recentOrders: Order[];
  hourlyBreakdown: HourlyCount[];
}

export interface StallSummary {
  stallId: string;
  stallName: string;
  mallName: string;
  slug: string;
  ordersToday: number;
  avgWaitTimeMinutes: number;
  activeOrders: number;
}

export interface AdminAnalytics {
  stalls: StallSummary[];
  totalOrdersToday: number;
  totalStalls: number;
}

export interface ApiError {
  error: string;
  message: string;
}

export type ListOrdersByStallParams = {
  status?: ListOrdersByStallStatus;
};

export type ListOrdersByStallStatus =
  (typeof ListOrdersByStallStatus)[keyof typeof ListOrdersByStallStatus];

export const ListOrdersByStallStatus = {
  waiting: "waiting",
  ready: "ready",
  completed: "completed",
} as const;

export type GetAdminAnalyticsParams = {
  password: string;
};