# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OrderPing - Food court order notification system API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stalls
    description: Stall operations
  - name: orders
    description: Order operations
  - name: analytics
    description: Analytics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stalls:
    get:
      operationId: listStalls
      tags: [stalls]
      summary: List all stalls
      responses:
        "200":
          description: List of stalls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stall"
    post:
      operationId: createStall
      tags: [stalls]
      summary: Register a new stall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStallBody"
      responses:
        "201":
          description: Stall created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stall"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stalls/{slug}:
    get:
      operationId: getStallBySlug
      tags: [stalls]
      summary: Get stall by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stall details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stall"
        "404":
          description: Stall not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stalls/{slug}/verify-password:
    post:
      operationId: verifyStallPassword
      tags: [stalls]
      summary: Verify vendor password
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPasswordBody"
      responses:
        "200":
          description: Password verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPasswordResult"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Stall not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stalls/{slug}/orders:
    get:
      operationId: listOrdersByStall
      tags: [orders]
      summary: List orders for a stall
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [waiting, ready, completed]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create/register an order (customer enters token)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stalls/{slug}/orders/{receiptNumber}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status by receipt number
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: receiptNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stalls/{slug}/orders/{receiptNumber}/ready:
    post:
      operationId: markOrderReady
      tags: [orders]
      summary: Mark order as ready (vendor action)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: receiptNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order marked ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stalls/{slug}/orders/{receiptNumber}/complete:
    post:
      operationId: markOrderCompleted
      tags: [orders]
      summary: Mark order as completed/picked up
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: receiptNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order marked completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /stalls/{slug}/orders/{receiptNumber}/nudge:
    post:
      operationId: nudgeCustomer
      tags: [orders]
      summary: Nudge customer to collect their order
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: receiptNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Nudge sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /stalls/{slug}/queue-status:
    get:
      operationId: getQueueStatus
      tags: [orders]
      summary: Get live queue status for a stall
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStatus"

  /stalls/{slug}/analytics:
    get:
      operationId: getStallAnalytics
      tags: [analytics]
      summary: Get analytics for a stall
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stall analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StallAnalytics"

  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [analytics]
      summary: Get analytics for all stalls (admin)
      parameters:
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Stall:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mallName:
          type: string
        slug:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        currentlyServing:
          type: string
          nullable: true
      required:
        - id
        - name
        - mallName
        - slug
        - email
        - createdAt

    CreateStallBody:
      type: object
      properties:
        name:
          type: string
        mallName:
          type: string
        email:
          type: string
        password:
          type: string
        slug:
          type: string
      required:
        - name
        - mallName
        - email
        - password

    VerifyPasswordBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    VerifyPasswordResult:
      type: object
      properties:
        success:
          type: boolean
        stallId:
          type: string
        stallName:
          type: string
      required:
        - success
        - stallId
        - stallName

    Order:
      type: object
      properties:
        id:
          type: string
        stallId:
          type: string
        stallSlug:
          type: string
        receiptNumber:
          type: string
        status:
          type: string
          enum: [waiting, ready, completed]
        createdAt:
          type: string
          format: date-time
        readyAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        nudgeCount:
          type: integer
        lastNudgeAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - stallId
        - stallSlug
        - receiptNumber
        - status
        - createdAt
        - nudgeCount

    CreateOrderBody:
      type: object
      properties:
        receiptNumber:
          type: string
      required:
        - receiptNumber

    QueueStatus:
      type: object
      properties:
        stallName:
          type: string
        mallName:
          type: string
        currentlyServing:
          type: string
          nullable: true
        waitingCount:
          type: integer
        recentlyServed:
          type: array
          items:
            type: string
      required:
        - stallName
        - mallName
        - waitingCount
        - recentlyServed

    StallAnalytics:
      type: object
      properties:
        ordersToday:
          type: integer
        avgWaitTimeMinutes:
          type: number
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        hourlyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/HourlyCount"
      required:
        - ordersToday
        - avgWaitTimeMinutes
        - recentOrders
        - hourlyBreakdown

    HourlyCount:
      type: object
      properties:
        hour:
          type: string
        count:
          type: integer
      required:
        - hour
        - count

    AdminAnalytics:
      type: object
      properties:
        stalls:
          type: array
          items:
            $ref: "#/components/schemas/StallSummary"
        totalOrdersToday:
          type: integer
        totalStalls:
          type: integer
      required:
        - stalls
        - totalOrdersToday
        - totalStalls

    StallSummary:
      type: object
      properties:
        stallId:
          type: string
        stallName:
          type: string
        mallName:
          type: string
        slug:
          type: string
        ordersToday:
          type: integer
        avgWaitTimeMinutes:
          type: number
        activeOrders:
          type: integer
      required:
        - stallId
        - stallName
        - mallName
        - slug
        - ordersToday
        - avgWaitTimeMinutes
        - activeOrders

    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message