# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OrderPing - Food court order notification system API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all stalls
 */
export const ListStallsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  mallName: zod.string(),
  slug: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
  currentlyServing: zod.string().nullish(),
});
export const ListStallsResponse = zod.array(ListStallsResponseItem);

/**
 * @summary Register a new stall
 */
export const CreateStallBody = zod.object({
  name: zod.string(),
  mallName: zod.string(),
  email: zod.string(),
  password: zod.string(),
  slug: zod.string().optional(),
});

/**
 * @summary Get stall by slug
 */
export const GetStallBySlugParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetStallBySlugResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  mallName: zod.string(),
  slug: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
  currentlyServing: zod.string().nullish(),
});

/**
 * @summary Verify vendor password
 */
export const VerifyStallPasswordParams = zod.object({
  slug: zod.coerce.string(),
});

export const VerifyStallPasswordBody = zod.object({
  password: zod.string(),
});

export const VerifyStallPasswordResponse = zod.object({
  success: zod.boolean(),
  stallId: zod.string(),
  stallName: zod.string(),
});

/**
 * @summary List orders for a stall
 */
export const ListOrdersByStallParams = zod.object({
  slug: zod.coerce.string(),
});

export const ListOrdersByStallQueryParams = zod.object({
  status: zod.enum(["waiting", "ready", "completed"]).optional(),
});

export const ListOrdersByStallResponseItem = zod.object({
  id: zod.string(),
  stallId: zod.string(),
  stallSlug: zod.string(),
  receiptNumber: zod.string(),
  status: zod.enum(["waiting", "ready", "completed"]),
  createdAt: zod.coerce.date(),
  readyAt: zod.coerce.date().nullish(),
  completedAt: zod.coerce.date().nullish(),
  nudgeCount: zod.number(),
  lastNudgeAt: zod.coerce.date().nullish(),
});
export const ListOrdersByStallResponse = zod.array(
  ListOrdersByStallResponseItem,
);

/**
 * @summary Create/register an order (customer enters token)
 */
export const CreateOrderParams = zod.object({
  slug: zod.coerce.string(),
});

export const CreateOrderBody = zod.object({
  receiptNumber: zod.string(),
});

/**
 * @summary Get order status by receipt number
 */
export const GetOrderParams = zod.object({
  slug: zod.coerce.string(),
  receiptNumber: zod.coerce.string(),
});

export const GetOrderResponse = zod.object({
  id: zod.string(),
  stallId: zod.string(),
  stallSlug: zod.string(),
  receiptNumber: zod.string(),
  status: zod.enum(["waiting", "ready", "completed"]),
  createdAt: zod.coerce.date(),
  readyAt: zod.coerce.date().nullish(),
  completedAt: zod.coerce.date().nullish(),
  nudgeCount: zod.number(),
  lastNudgeAt: zod.coerce.date().nullish(),
});

/**
 * @summary Mark order as ready (vendor action)
 */
export const MarkOrderReadyParams = zod.object({
  slug: zod.coerce.string(),
  receiptNumber: zod.coerce.string(),
});

export const MarkOrderReadyResponse = zod.object({
  id: zod.string(),
  stallId: zod.string(),
  stallSlug: zod.string(),
  receiptNumber: zod.string(),
  status: zod.enum(["waiting", "ready", "completed"]),
  createdAt: zod.coerce.date(),
  readyAt: zod.coerce.date().nullish(),
  completedAt: zod.coerce.date().nullish(),
  nudgeCount: zod.number(),
  lastNudgeAt: zod.coerce.date().nullish(),
});

/**
 * @summary Mark order as completed/picked up
 */
export const MarkOrderCompletedParams = zod.object({
  slug: zod.coerce.string(),
  receiptNumber: zod.coerce.string(),
});

export const MarkOrderCompletedResponse = zod.object({
  id: zod.string(),
  stallId: zod.string(),
  stallSlug: zod.string(),
  receiptNumber: zod.string(),
  status: zod.enum(["waiting", "ready", "completed"]),
  createdAt: zod.coerce.date(),
  readyAt: zod.coerce.date().nullish(),
  completedAt: zod.coerce.date().nullish(),
  nudgeCount: zod.number(),
  lastNudgeAt: zod.coerce.date().nullish(),
});

/**
 * @summary Nudge customer to collect their order
 */
export const NudgeCustomerParams = zod.object({
  slug: zod.coerce.string(),
  receiptNumber: zod.coerce.string(),
});

export const NudgeCustomerResponse = zod.object({
  id: zod.string(),
  stallId: zod.string(),
  stallSlug: zod.string(),
  receiptNumber: zod.string(),
  status: zod.enum(["waiting", "ready", "completed"]),
  createdAt: zod.coerce.date(),
  readyAt: zod.coerce.date().nullish(),
  completedAt: zod.coerce.date().nullish(),
  nudgeCount: zod.number(),
  lastNudgeAt: zod.coerce.date().nullish(),
});

/**
 * @summary Get live queue status for a stall
 */
export const GetQueueStatusParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetQueueStatusResponse = zod.object({
  stallName: zod.string(),
  mallName: zod.string(),
  currentlyServing: zod.string().nullish(),
  waitingCount: zod.number(),
  recentlyServed: zod.array(zod.string()),
});

/**
 * @summary Get analytics for a stall
 */
export const GetStallAnalyticsParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetStallAnalyticsResponse = zod.object({
  ordersToday: zod.number(),
  avgWaitTimeMinutes: zod.number(),
  recentOrders: zod.array(
    zod.object({
      id: zod.string(),
      stallId: zod.string(),
      stallSlug: zod.string(),
      receiptNumber: zod.string(),
      status: zod.enum(["waiting", "ready", "completed"]),
      createdAt: zod.coerce.date(),
      readyAt: zod.coerce.date().nullish(),
      completedAt: zod.coerce.date().nullish(),
      nudgeCount: zod.number(),
      lastNudgeAt: zod.coerce.date().nullish(),
    }),
  ),
  hourlyBreakdown: zod.array(
    zod.object({
      hour: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get analytics for all stalls (admin)
 */
export const GetAdminAnalyticsQueryParams = zod.object({
  password: zod.coerce.string(),
});

export const GetAdminAnalyticsResponse = zod.object({
  stalls: zod.array(
    zod.object({
      stallId: zod.string(),
      stallName: zod.string(),
      mallName: zod.string(),
      slug: zod.string(),
      ordersToday: zod.number(),
      avgWaitTimeMinutes: zod.number(),
      activeOrders: zod.number(),
    }),
  ),
  totalOrdersToday: zod.number(),
  totalStalls: zod.number(),
});